car_state/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
ultralytics>=8.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6